fix: qubit index 转 python int, 修复 pyscf 返回 numpy int 时 tensorcircuit 崩
pyscf 某些版本 (如 2.7) 的 mol.nao_nr() 返回 numpy int 而非 python int。
tensorcircuit 0.12 的 apply_list 对 numpy int 类型的 qubit index 包装崩成
空结构 (for-else 的 else 分支), 抛 "Illegal index specification"。
表现为 c.x(0) 正常但 c.x(norb) 崩 (类型差异, 非值差异)。

修复 (3 处 int() 防御):
- src/tc_sqd/lucj.py _qubit: 返回 int(...)
- examples/h2_sqd_demo.py: norb = int(mol.nao_nr())
- tests/test_h2_sqd.py make_tc_circuit: norb = int(norb)

验证: 故意传 np.int64 给 build_lucj_circuit 不再崩 (bsm 正常采样);
tc312 环境 (pyscf 2.14) tests 9/9 与 demo 均不回归。

+++ src/tc_sqd/lucj.py
@@ -74,9 +74,11 @@ def _qubit(spin: str, orb: int, norb: int) -> int:
 
     α_i → 2norb-1-i ; β_i → norb-1-i  (q0=MSB, [β|α] 布局)。
     """
+    # int() 防御: pyscf 某些版本 (如 2.7) mol.nao_nr() 返回 numpy int,
+    # 传给 tensorcircuit 的 c.x(index) 会触发 "Illegal index specification"
     if spin == "a":
-        return 2 * norb - 1 - orb
-    return norb - 1 - orb
+        return int(2 * norb - 1 - orb)
+    return int(norb - 1 - orb)
 
 
 def build_lucj_circuit(
